fix issues with log record with error
